[Bronze IV] Title: Petrol, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Bronze/18330. Petrol/Petrol.cc
@@ -4,7 +4,7 @@ using std::cout;
 
 int main()
 {
-    long int n, k;
+    int n, k;
     cin >> n >> k;
     cout << ((n <= (60 + k)) ? n * 1500 : (((60 + k) * 1500) + ((n - (60 + k)) * 3000)));
     return 0;
